Separated key rotation and key generation logic
