+++ src/index.js
@@ -2,5 +2,6 @@
 
 module.exports = {
 	KeyGenerator: require('./key-generator'),
+	CoreKeyGenerator: require('./core-key-generator'),
 	AbstractPublicKeyStore: require('./abstract-public-key-store')
 };
